pi_service/requirements.txt
black
RPi.GPIO
PyYAML